DataCollection/requirements.txt
influxdb_client==1.45.0
matplotlib==3.9.2
neurosity==2.2.0
pandas==2.2.2
python-dotenv==1.0.1
tk
loguru
sqlalchemy
psycopg2
pandas